Remote Script Executor
To use this script simply replace "C:\Directory\ScriptsDirectory" with the directory that your scripts are located, replace "C:\Directory\ListOfServers.txt" with the full path to your list of servers that you would like to run the script on (separated by a line break per machine), and replace "domain\username" with your domain and username. It will then ask you to enter the number script you would like to run. It will prompt to enter your credentials and then execute the script using your credentials.