Query-Baselines.psm1
To be used on a system that has already ran Create-Baselines. This script queries the services on a system and compares them to the baseline csv file that is created after running Create-Baselines. By setting $CorrectServices to true, the script will also set the services that are currently in the incorrect service state to the correct state. Setting $LogBadServices to true will log the delinquent services to the text file "C:\Maintenance\Baselines\BadServices.txt".